Refactor machine process PoC structure
