+++ machine-process-visibility-poc/backend/constants.py
@@ -1,0 +1,4 @@
+DESCRIPTION_TEMPLATE = "【状態】\n\n【注意】\n\n【次工程】\n"
+PROCESSES = ["未振り分け", "内面研磨", "刃物研磨", "機械加工", "板金加工", "手加工", "完了"]
+COMMENT_TYPES = {"作業", "気づき", "異常", "補足"}
+DEFAULT_PASSWORD = "password"
